+++ drivers/char/ftape/kernel-interface.c
@@ -61,10 +61,10 @@ static int ftape_open(struct inode *ino, struct file *filep);
 static void ftape_close(struct inode *ino, struct file *filep);
 static int ftape_ioctl(struct inode *ino, struct file *filep,
 		       unsigned int command, unsigned long arg);
-static int ftape_read(struct inode *ino, struct file *fp, char *buff,
-		      int req_len);
-static int ftape_write(struct inode *ino, struct file *fp, const char *buff,
-		       int req_len);
+static long ftape_read(struct inode *ino, struct file *fp,
+		       char *buff, unsigned long req_len);
+static long ftape_write(struct inode *ino, struct file *fp,
+			const char *buff, unsigned long req_len);
 
 static struct file_operations ftape_cdev =
 {
@@ -310,13 +310,14 @@ static int ftape_ioctl(struct inode *ino, struct file *filep,
 
 /*      Read from tape device
  */
-static int ftape_read(struct inode *ino, struct file *fp, char *buff, int req_len)
+static long ftape_read(struct inode *ino, struct file *fp,
+	char *buff, unsigned long req_len)
 {
 	TRACE_FUN(5, "ftape_read");
 	int result = -EIO;
 	int old_sigmask;
 
-	TRACEi(5, "called with count:", req_len);
+	TRACEi(5, "called with count:", (int) req_len);
 	if (!busy_flag || MINOR(ino->i_rdev) != ftape_unit || ftape_failure) {
 		TRACE(1, "failed: not busy, failure or wrong unit");
 		TRACE_EXIT;
@@ -333,13 +334,14 @@ static int ftape_read(struct inode *ino, struct file *fp, char *buff, int req_le
 
 /*      Write to tape device
  */
-static int ftape_write(struct inode *ino, struct file *fp, const char *buff, int req_len)
+static long ftape_write(struct inode *ino, struct file *fp,
+	const char *buff, unsigned long req_len)
 {
 	TRACE_FUN(8, "ftape_write");
 	int result = -EIO;
 	int old_sigmask;
 
-	TRACEi(5, "called with count:", req_len);
+	TRACEi(5, "called with count:", (int) req_len);
 	if (!busy_flag || MINOR(ino->i_rdev) != ftape_unit || ftape_failure) {
 		TRACE(1, "failed: not busy, failure or wrong unit");
 		TRACE_EXIT;
