+++ fs/proc/kmsg.c
@@ -28,7 +28,8 @@ static void kmsg_release(struct inode * inode, struct file * file)
 	(void) sys_syslog(0,NULL,0);
 }
 
-static int kmsg_read(struct inode * inode, struct file * file,char * buf, int count)
+static long kmsg_read(struct inode * inode, struct file * file,
+	char * buf, unsigned long count)
 {
 	return sys_syslog(2,buf,count);
 }
