+++ fs/sysv/dir.c
@@ -21,7 +21,8 @@
 
 #include <asm/segment.h>
 
-static int sysv_dir_read(struct inode * inode, struct file * filp, char * buf, int count)
+static long sysv_dir_read(struct inode * inode, struct file * filp,
+	char * buf, unsigned long count)
 {
 	return -EISDIR;
 }
