+++ fs/sysv/file.c
@@ -33,7 +33,7 @@
 #include <linux/fs.h>
 #include <linux/sysv_fs.h>
 
-static int sysv_file_write(struct inode *, struct file *, const char *, int);
+static long sysv_file_write(struct inode *, struct file *, const char *, unsigned long);
 
 /*
  * We have mostly NULL's here: the current defaults are ok for
@@ -72,7 +72,8 @@ struct inode_operations sysv_file_inode_operations = {
 	NULL			/* permission */
 };
 
-int sysv_file_read(struct inode * inode, struct file * filp, char * buf, int count)
+long sysv_file_read(struct inode * inode, struct file * filp,
+	char * buf, unsigned long count)
 {
 	struct super_block * sb = inode->i_sb;
 	int read,left,chars;
@@ -199,7 +200,8 @@ int sysv_file_read(struct inode * inode, struct file * filp, char * buf, int cou
 	return read;
 }
 
-static int sysv_file_write(struct inode * inode, struct file * filp, const char * buf, int count)
+static long sysv_file_write(struct inode * inode, struct file * filp,
+	const char * buf, unsigned long count)
 {
 	struct super_block * sb = inode->i_sb;
 	off_t pos;
